V2.28 V2.4 milestone (#258)
* #85 add KORCore class libary

* #85 add core models and utility

* #85 add database notify changed

* #85 update ignore

* #84 #85

* #82 #85

* #82 #85 fix DI

* #81 #82 #85

* #81 #82 #85 fix android build error

* #81 #82 #85

* #81 #82  #85 add SignalR

* #81 #82 #85

* #81 #82 #85 Alpha 1

* 85 2 tracking database changes real time (#229)

* #85 add KORCore class libary

* #85 add core models and utility

* #85 add database notify changed

* #85 update ignore

* #84 #85

* #82 #85

* #82 #85 fix DI

* #81 #82 #85

* #81 #82 #85 fix android build error

* #81 #82 #85

* #81 #82  #85 add SignalR

* #81 #82 #85

* #81 #82 #85 Alpha 1

* #81 #82 #85 fix color

* #81 # 82 #85 add error

* #0

* Update

* #81 #82 #85

* #86

* #86 add error log

* #86 add mainthred

* #86 fix color

* Fix setting page-error, add kncsk_tool (#231)

* 85 2 tracking database changes real time (#232)

* Fix setting page-error, add kncsk_tool

* Fix github actions

* 85 2 tracking database changes real time (#233)

* Fix setting page-error, add kncsk_tool

* Fix github actions

* Change action to GitHub runner

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml (#234)

* Update dev.yaml (#235)

* Update publish.yaml (#237)

* Update dev.yaml (#239)

* Update dev.yaml (#240)

* Update dev.yaml (#241)

* #242 add about page, update translate (#245)

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update dev.yaml

* Update KoOrderRegister.csproj

* Update KoOrderRegister.csproj (#246)

* Update AppRes.de-DE.resx (#247)

* Update AppRes.fr-FR.resx (#248)

* Update dev.yaml (#249)

* Update dev.yaml (#250)

* Create release_base.md (#251)

* Update dev.yaml (#252)

* Update publish.yaml (#253)

* Update publish.yaml (#254)

* 244 add russian language (#255)

* #244 add russion language and machin translator

* #244 add rusian language, translator, workflows

#244 add rusian language and add machin translator python script and update workflows

* update workflow

* #243 update settings page UI (#256)

* #fix bug

* Create readme.md (#257)

+++ KoOrderRegister/Modules/Order/ViewModels/OrderListViewModel.cs
@@ -1,7 +1,5 @@
 ﻿using KoOrderRegister.Localization;
 using KoOrderRegister.Localization.SupportedLanguage;
-using KoOrderRegister.Modules.Database.Models;
-using KoOrderRegister.Modules.Database.Services;
 using KoOrderRegister.Modules.Order.Pages;
 using KoOrderRegister.Services;
 using KoOrderRegister.Utility;
@@ -15,20 +13,24 @@ using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Input;
+using KORCore.Modules.Database.Models;
+using KORCore.Modules.Database.Services;
+using KORCore.Utility;
+using KORCore.Modules.Database.Factory;
 
 namespace KoOrderRegister.Modules.Order.ViewModels
 {
     public class OrderListViewModel : BaseViewModel
     {
-        private readonly IDatabaseModel _database;
+        private IDatabaseModel _database;
         private readonly OrderDetailsPage _orderDetailsPage;
+        private readonly IDatabaseModelFactory _databaseModelFactory;
         #region Binding varrible
-       
+
         #endregion
         public string SearchTXT { get; set; } = string.Empty;
-        private CancellationTokenSource _cancellationTokenSource;
+        private CancellationTokenSource _cancellationTokenSource = new CancellationTokenSource();
 
- 
         public ObservableCollection<OrderModel> Orders { get; set; } = new ObservableCollection<OrderModel>();
 
         #region Commands
@@ -39,9 +41,10 @@ namespace KoOrderRegister.Modules.Order.ViewModels
         public ICommand SearchCommand { get; }
         #endregion
 
-        public OrderListViewModel(IDatabaseModel database, OrderDetailsPage orderDetailsPage, IAppUpdateService updateService, ILocalNotificationService notificationService) : base(updateService, notificationService)
+        public OrderListViewModel(IDatabaseModelFactory database, OrderDetailsPage orderDetailsPage, IAppUpdateService updateService, ILocalNotificationService notificationService) : base(updateService, notificationService)
         {
-            _database = database;
+            _databaseModelFactory = database;
+            _database = database.Get();
             _orderDetailsPage = orderDetailsPage;
 
             AddNewOrderCommand = new Command(NewOrder);
@@ -51,29 +54,26 @@ namespace KoOrderRegister.Modules.Order.ViewModels
             SearchCommand = new Command<string>(Search);
         }
 
-        public event PropertyChangedEventHandler PropertyChanged;
-
-        protected virtual void OnPropertyChanged(string propertyName = null)
+        public override void OnAppearing()
         {
-            try
-            {
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-            }
-            catch (TargetInvocationException ex)
-            {
-                Debug.WriteLine($"Inner Exception: {ex.InnerException}");
-            }
+            _database = _databaseModelFactory.Get();
         }
 
         public async void NewOrder()
         {
-            await App.Current.MainPage.Navigation.PushAsync(_orderDetailsPage);
+            if (App.Current?.MainPage != null)
+            {
+                await App.Current.MainPage.Navigation.PushAsync(_orderDetailsPage);
+            }
         }
 
         public async void EditOrder(OrderModel order)
         {
-            _orderDetailsPage.EditOrder(order);
-            await App.Current.MainPage.Navigation.PushAsync(_orderDetailsPage);
+            if (App.Current?.MainPage != null)
+            {
+                _orderDetailsPage.EditOrder(order);
+                await App.Current.MainPage.Navigation.PushAsync(_orderDetailsPage);
+            }
         }
 
         public async void UpdateOrders()
@@ -99,6 +99,7 @@ namespace KoOrderRegister.Modules.Order.ViewModels
                 }
             }
         }
+
         private async Task _updateOrders()
         {
             IsRefreshing = true;
@@ -127,16 +128,19 @@ namespace KoOrderRegister.Modules.Order.ViewModels
 
         public async void DeleteOrder(OrderModel order)
         {
-            bool result = await Application.Current.MainPage.DisplayAlert(AppRes.Delete, AppRes.AreYouSureYouWantToDelete + " " + order.OrderNumber, AppRes.Yes, AppRes.No);
-            if (result)
+            if (Application.Current?.MainPage != null)
             {
-                if (await _database.DeleteOrder(order.Guid) > 0)
+                bool result = await Application.Current.MainPage.DisplayAlert(AppRes.Delete, AppRes.AreYouSureYouWantToDelete + " " + order.OrderNumber, AppRes.Yes, AppRes.No);
+                if (result)
                 {
-                    Orders.Remove(order);
-                }
-                else
-                {
-                    await Application.Current.MainPage.DisplayAlert(AppRes.Delete, AppRes.FailedToDelete + " " + order.OrderNumber, AppRes.Ok);
+                    if (await _database.DeleteOrder(order.Guid) > 0)
+                    {
+                        Orders.Remove(order);
+                    }
+                    else
+                    {
+                        await Application.Current.MainPage.DisplayAlert(AppRes.Delete, AppRes.FailedToDelete + " " + order.OrderNumber, AppRes.Ok);
+                    }
                 }
             }
         }
